requirements.txt
flask
python-dotenv
openai-whisper
transformers
sentencepiece
torch
moviepy
edge-tts
demucs
openai